Moved astar.js to lib
