+++ js/lib/astar.js
